backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-multipart>=0.0.6
pillow>=10.0.0
pytesseract>=0.3.10
phonenumbers>=8.13.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
